requirements.txt
loguru
RPi.GPIO